requirements.txt
pillow
tkcap